requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
dateparser==1.1.1
idna==3.3
python-dateutil==2.8.2
python-kucoin==2.1.3
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
requests==2.28.1
six==1.16.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.10
websockets==10.3
